Make Symbols-Only webfonts available stand alone
[why]
When people want to use the Symbols-Only webfont that we use for the
website directly (without hosting the css/woff files themselves) they
probably need a stub to access our css.

I have no clue what consequences this has for our traffic, and probably
we need to remove this if problems come up.

Origin of this is a request via gitter:

Mikle-Bond (Michael)
  Hello! I have a request about webfont. Can you make the
  nerd-fonts-generated{,.min}.css files available via
  https://www.nerdfonts.com/assets/css/ ? So that users could
  reference them directly and have SymbolsOnly variant of the
  font with classes names.

  I have found this [not linked here] message from a month ago,
  [in fact 2 years and 1 month] and my request could help that
  person too. They were having problem, as nerd-fonts-generated.css
  is only available via githubusercontent.com at the moment, and
  it doesn't point correctly to the font file. But it does so for
  nerdfonts.com website, after it gets bundled into combo.css.

[how]
Make nerd-fonts-generated.min.css availabe directly as
https://www.nerdfonts.com/assets/css/webfont.css.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>