Remove Hack postprocessing
[why]
The special Hack postprocessing does this:
 * Add a dummy signature
 * Set fsType to zero
 * run ttfautohint on the font

There once has been reason to do this (probably), but on currently
supported fontforges this is not needed anymore.

 * Fonts run fine without the signature
 * fsType is anyhow zero, set by fontforge
 * the preexisting hinting is kept by fontforge
   (furthermore: We do not want hints for the symbols usually)

Double-checked source and patched font with ttx, instructions (hints)
are definitively existing and equal in the patched font.

So we do not need any of this special postprocessing anymore.

See also commit
  e9d84b6c0  Adds WIP scripts and logic for post processing patched fonts (WIP for fixing #70)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>