iA-Writer: Remove blanks in pathnames
[why]
Of course it is nice to have blanks in the pathnames, because that means
the filename handling has to be very correct everywhere, quoting the
names and so on.
But it is also a nuisance. And it does not look too good on the command
line.

[how]
Like Meslo, JetBrains Mono, Noto, and Overpass we do not repeat the base
font name in the directory name. (Counterexample MPlus...)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>