Drop Symbols 1000 em font
[why]
There is really no reason to have two identical fonts that just have a
different EM.

The family name changed to the previous release from
"Symbols Nerd Font" to "Symbols2048Em Nerd Font"

[why]
As this is a ttf font we choose to keep 2048, and drop the 1000.
When we drop the number the family name is also correct again.
The family name is used by the fontconfig that people probably have
installed, so we should not change that familyname.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>