font-patcher: Fix logging related warning
[why]
On some fonts the patcher tries to raise a warning, but it can not.

[how]
The 'overlap' variable can hold a floating point number or be None (if
there shall be no overlap processing. Formerly we output that value
using the {} format, but the logging module does not have that. To
accomodate to different types use repr() instead.

[note]
Also improve some other logging related calls.

Reported-by: Jim Myhrberg <contact@jimeh.me>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>