Update cheat sheet WITH removed icons
Update the cheat sheet to v3.0.0 but also show the removed icons of
v2.3.3. That style is .nfold (read: nf-old) and it is just hacked in.
I believe it can be removed later on, with 3.1.0 or something at least.

We also need the old (removed) icons as a font to display them despite
being removed. A small helper script 'archives' just the dropped glyphs
into a new webfont.
The script can be removed afterwards, this is a single run, archiving
the dropped after v2.3.3 icons.

See pairing commit in master branch:
    Date:   Wed May 3 17:54:44 2023 +0200
    generate-css: Create cheat sheet with removed icons

[note]
Also remove unused `nerd-fonts-generated.css`, wich confuses me every
time. It's sufficient to have it in the master branch for people, we do
not use it here.

f