CascadiaCode: Correct version string
[why]
With commit
  b6301e590  CascadiaCode: Rehint and use ttf

we replaced the hints done upstream by ttfautohint with hints done via
Visual TrueType, because they render the text more similar to the
variable fonts (hints).

But the version string still mentions 'ttfautohint' because we did not
touch the font files at all except opening, hinting, closing in VTT.
This can be confusing.

[how]
Change the version string to "Version 2111.001; VTT 6.35" via:

* `ttx -i *.ttf`
* Edit `ttx` files and replace version, see above
* `ttx -b *.ttx`
* Move new fonts over old fonts

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>