Fix changelog formatting
I think the `</div> </details>` at the end of the v3.0.1 release notes was being parsed as part of the last `<li>` item, resulting in `</div> </details>` being escaped. This commit should fix the issue where in the HTML output, the v3.0.1 release `<details>` included the release notes for v3.0.0, v2.3.3, etc.