cheat-sheet: Allow incomplete searches
[why]
Sometimes one does not know what the exact search term is.
For example if you want to find a 'homefolder' but the name is
'homedirectory' it is impossible to find.

[how]
Allow prefix search, in this case at least 'home' will find both
variants.
I do believe a full substring search would be even better, but that is
not supported by minisearch.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>