docker: Re-add "master" tag
[why]
Previously we had the tag 'master' set in all images, as the images were
created with every push the default branch (i.e. master). Some people
might use that tag for something.
At least we can not remove the last image with 'master' tag, and it will
become outdated immediately, making it a misnomer.

[how]
Add both tags. Hope this works ;-)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>