cheat-sheet: Show no icons by default
[why]
The additional hints in the bottom are hard to notice if we flood users
with all available items first.
Usually users want to search the icons and not browse, so showing them
all per default seems better.

[how]
Just do not change the search term from "" to something else, so the
search comes up empty on an empty search.

If people search for a blank the previous mechanics to show all icons is
triggered instead. A message to hint for this is added to the empty
search result.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>