Allow searching for pasted icon
[why]
There are cases that I find myself into sometimes when I have the glyph
with me but not its name or codepoint. It would be nice to be able to
search for the codepoint and/or name of a glyph by literally putting it
into the search box of the cheat-sheet.
There are several reasons to know the name of the glyph, to use it in a
webpage, to look for alternatives when they get removed, etc.

[how]
Convert high codepoint-chars to codepoint text.

Fixes: #1307

Authored-by: ad-chaos
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>