Fix navbar in two lines
[why]
The navbar should only be one line, switching to smaller and smaller
items, until we reach a very small size (the smallest) where we finally
allow it to overlow into the next lines.
But that is not the case, the navbar switches between one and two lines
already at medium sized screens.

[how]
Correct the width point when the navbar titles switch from long-titles
to short-titles.

Introduce new width switch point that reduces the cell widths of the navbar
one they become very full and further reduction in width would result in
overflow into the next line.

Tested on Netscape and Chromium; the later needs bigger numbers (hmm)

Fixes: #1375

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>