Fix Star/Fork button sometimes not showing
[why]
The Star and Fork buttons sometime show nicely, sometimes they are just
plain linked tect with no count and no icon and nothing.

[how]
The reason is not completely clear, sometimes it works and without any
change after a complete reload it does not work anymore.

The most likely candidate is the Github rate limit that we run into for
cross repo access. To avoid that we just copy the button script into our
repository. The rate limit will now not apply anymore.

But of course we will also not get any updates. That needs to be checked
from time to time manually - as with the Github actions.

Strange is already that the current script from github.io is version
2.27.0 while the latest release on the repo is 4.2.2. The date is the
same (Dec 2022).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>