Redo: Fix wrong small version of all-icons-divider
[why]
The previous fix broke the instances of the subtitle circle in all the
other topics (Downloads, CheatSheet, Backers, ...) because it rendered
the circle too wide horizontally, while the vertical size was ok.

[how]
The previous previous fix with a concrete circle hight and width and
adequate line spacing for the filling font works better. The circle is
not absolutely circular, but roughly it is.
Unless there is a Web magician who can fix this in a better way I
believe the not-really-circular in one place (all-icons-divider) is
better than absurdly wide circles in all section subtitles.

Fixes: #1384

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>