Martian: Tweak naming
[why]
With Martian Mono we have a font that uses aggressive abbreviation for
the weights and styles (two letters mostly) in all name fields. That is
very hard to parse.

[how]
We rewrite the filenames such that it contains all the necessary information
and use the new patcher flag use the filename as naming source.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>