Cascadia Mono: Rehint
CascadiaCode: Rehint and use ttf

[why]
A lot people (read: People on Windows) have the variable font (VF) version of
Cascadia Mono installed - it comes bundled with Windows Terminal.

The static Cascadia Mono instances that Microsoft releases are hinted
with ttfautohint which creates small sized glyphs that are visibly very
different. People compare the static Caskaydia Mono with the VF Cascadia
Mono and are surprised.

[how]
We need to re-hint all the fonts, to get hints that are comparable
to the VF hints. We can not use the hints of the VF because the outlines
are different: The VF has (of course) overlapping outlines, while the
static ones (as usual) have not.

The re-hinting can be done with VTT or TTH - both showed results that
are more like the original VF font. The usual ttfautohint has been used
of the static fonts in the font release and can not be used. It is the
reason for this whole problem.

* Used VTT 6.35
* Open font file in VTT
* Import all programs
* Generate 'VTT talk' via Tools -> AutoHint -> LightLatinAutoHint
* Save font file as ...

References:
https://github.com/microsoft/cascadia-code/issues/371
https://learn.microsoft.com/en-us/typography/tools/vtt/

See also commit
  b6301e5 CascadiaCode: Rehint and use ttf

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>