Add small tool to create release artefact table
Usage is a bit ... hard, because we need to do this before the release
is published.

Download the complete artifact zip archive from the release-drafting
run.

Unpack it into the repository (will end up in patched-fonts/)

Pack all archives with archive-fonts.sh

Then use this tool to extract the names and sizes of the artifacts.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>