NotoSans: Update to 2.013
All the different ligature removal calls are becoming more and more a
problem, because we have only one config.json for all fonts in a
font-root-directory. But the Noto variants have the ligatures in
different lookups. With the new font we finally remove lookups in the
new font that have a completely different contents then the lookups we
wanted to remove in the other family.

So the config.json is becoming more flexible: Now first we seach in the
concrete font file directory and only if there is no config we progress
to the font root directory and search there.

See also commit
d9f7dbe23  Prepatched fonts: Revive some ligature removal

Here we have
* fi and similar in lookup 41
* l-dot and similar in lookups 13, 14, 15

Fixes: #1472

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>