ComicShanns: Fix carons
[why]
The carons are all upside down.
The dcaron and tcaron should use an 'apothrophe'.

[how]
The simplify the fixing of all carons we take the excellent fork ofu
Rodrigo Batista de Moraes that (re?)creates the letters with caron from
references.

Just manually flip the caron and all combined glyphs will be fixed.

Create a combining-comma-above-right from the ordinary comma.

Reference that for dcaron and tcaron.

Rerun `make-bold` which creates the two otf files.

Fixes: #1492

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>