Revert fontconfig change
[why]
The fontconfig is autogenerated by the release process with the
bin/scripts/generate-fontconfig.sh script. The reason is that its
contents should match the released fonts (i.e. the latest release).

Also:
Upstream changed the name grouping a bit, so the superfluous "Bold"
family has been dropped (and does not need an entry anymore).

[how]
Revert the (manual) change and wait for the release process to
update the fontconfig.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>