name-parser: Fix keep-regular rule
[why]
The keep-regular rule should add 'Regular' as weight if there is no
weight (e.g. Bold) or style (e.g. Italic).

But when a font does have neither but has a widths specifier it is also
not added.

[how]
Check for some conrete weight and not only any entry in the weight
token, because we - unfortunately - mix weights and widths in that token
set.

We do not need to check the styles because it is technically impossible
for it to comtain Regular and something else.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>