Terminus: Remove old patched font directories
[why]
With commit
  23a14f184 Update Terminus to 4.49.3

we dropped the subdirectory structure of Terminus, but forgot to remove
that structure in the patched-fonts/ directory. The release will clear
the destination directory but not remove subdirectories, so we end up
with the 3.1.1 and 3.2.0 patched font files in the repo.

Luckily this is catched by the archiver process:

[Nerd Fonts]  Limiting archive to pattern 'Terminus'
[Nerd Fonts]  Packing Terminus.tar.xz (24 fonts)
[Nerd Fonts]  Did not pack expected number of font files! Likely same font names for different paths.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>