FontAwesome: Swap some icons
youtube and cloudsmith should change their codepoints, because youtube
occupies a 'lesser' youtube codepoint and yields a 'better' codepoint to
cloudsmith.

Better means the icon looks more like in FA 4.

Also swap the devices icons that in 4 showed a screen and now it is
filled.

[how]
The mechanics is a bit ... ugly.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>