FontAwesome: Leave a gap for Fira progress icons
[why]
We do not want to break the Fira progress icons, but on the other hand
these are so few icons that we can not keep the whole 0x0100 block
unoccupied.
So we just leave a small gap in the FontAwesome codepoints.

[how]
Be careful, the current code moves the icons 'to the back', which will
be different if we ever update FontAwesome. The code will break
(reassign a lot codepoints) then.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>